backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
python-dotenv==1.0.0
paramiko==2.12.0
python-dotenv==1.0.0
